almost there

+++ test/test_addition.py
@@ -25,7 +25,7 @@ class FirstStep(unittest.TestCase):
         self.assertEqual(ret[1,1], 2)
 
     def test_pyocl_add(self):
-        a = exprgraph.Variable(name='a', array=np.eye(3, 3))
+        a = exprgraph.Variable(name='a', array=np.eye(3, 3, dtype=np.float32))
         b = exprgraph.Variable(name='b', array=np.ones((3, 3), np.float32))
         expr = AddOP(a, b)
         f1 = function.function([expr], name='allonsy', platform='pyopencl')
